Rename some elements to better fit compose multiplatform theme
